Refactored api usage
