+++ ui/src/views/SignUp.vue
@@ -87,7 +87,7 @@
 </template>
 
 <script>
-	import authApi from '../auth/authApi';
+	import authApi from '../auth';
 
 	export default {
 		name: 'Signup',
